+++ applications/debug/ccid_test/iso7816_t0_apdu.h
@@ -26,11 +26,11 @@ void iso7816_answer_to_reset(Iso7816Atr* atr);
 void iso7816_read_command_apdu(
     struct ISO7816_Command_APDU* command,
     const uint8_t* dataBuffer,
-    uint32_t dataLen);
+    uint8_t dataLen);
 void iso7816_write_response_apdu(
     const struct ISO7816_Response_APDU* response,
     uint8_t* readerToPcDataBlock,
-    uint32_t* readerToPcDataBlockLen,
+    uint8_t* readerToPcDataBlockLen,
     uint8_t* responseDataBuffer,
-    uint32_t responseDataLen);
+    uint8_t responseDataLen);
 #endif //_ISO7816_T0_APDU_H_
